Changed search pattern for prokka from contents : 'organism:'  to fn: '*txt'
I changed search pattern for prokka from contents "organism:" to instead search for a file with the extension ".txt" since this is the suffix for all prokka summary statistics files. Unfortunately this is not very specific but I belive this will still be more efficient than searching all files in a folder structure such as the one prokka creates. I have also double checked with @fboulund that the prokka module is written so that it very quickly discards a file if it doesn't start like the one prokka creates, which should make the process as fast as possible.