#886 Fixing matplotlib plot bands
The default alignment for matplotlib plot bars is to be centered, but the relevant (x/y)PlotBands array is originally constructed (e.g. in the fastqc module) with the assumption that the bar should begin at the specified point. This fixes the broken plot bars when MultiQC auto-switches from highcharts to matplotlib.