Add option to rename lima samples
The lima output files only contain barcode names, and no sample
names. To allow the lima results to be properly integrated into the
broader MultiQC report, add an option --lima-barcodes to specify a file
with the mapping from barcodes to sample names.

When this is specified, the values in this file are used to update the
barcode names to sample names.