Update NanoStat search pattern
Use the string "General summary:" followed by 9 spaces in the first line
to distinguish a NanoStat report. Those 9 spaces probably aren't supposed
to be there, but they are found in all of the test files, and Linus says
it's a feature and not a bug once people start depending on it.

Signed-off-by: Thomas A. Christensen II <25492070+MillironX@users.noreply.github.com>