fix: clear accumulated non-N and total base counts between samples and mates
The existing implementation of the N content plot for dragen_fastqc incorrectly maintains counts
between samples and mates, causing the computed per-site N rate to be order-dependent (typically on
alphabetical ordering of input sample IDs), and to approach a cross-sample/read/lane/etc. average.
This isn't immediately apparent from output plots when the QCed libraries are performing well, but
becomes easily visible when one of the lanes, for example, has a large spike in N content.