Update hicpro.py
Previous code version just tried to convert the extracted value to integer with the 'int' func. If value is not an integer (eg 'NA', '34567.234e+09'), then the whole program could crash. 
The update first tries to convert to int, then to float. If both did not work, then the value is left as is.

Possible complications:
1) The code tries to convert to int at first, hence floats can be rounded.
2) This might theoretically also cause problems later (eg plots dont accept strings).