Update WgsMetrics.py
This resolves a off-by-one error in the histogram plot for this module. This issue is particularly problematic for shallow/low-pass sequencing with a high proportion of bases with zero coverage.

The problem is solved by reconfiguring the y-axis to plot 'the percentage of bases greater or equal than the current x-axis coverage' rather than 'the percentage of bases greater than the current x-axis coverage'