Lima: split delimiter is a tab, not any whitespace.
The file format is given as:

```
lima/counts:
  contents: "IdxFirst\tIdxCombined\tIdxFirstNamed\tIdxCombinedNamed\tCounts\tMeanScore"
```

Thus, tabs are indeed expected. We had a couple samples returned with "Not barcoded" as the sample ID. `.split()` of course is happy to split on that whitespace, and thus an invalid value was parsed for the counts. It must be split on a tab.