Custom logo: dark mode option, custom width option. (#3400)
Allow customisation of the custom logo - a variant image to use when the report is in dark mode, and the ability to specify a custom width.

Added a default width of 320px (the same as the MultiQC logo), which is now applied to the custom logo if no custom width is set.